Modify val type to unsigned int in icc_set_sgi
Modify val type from unsigned long to unsigned int in
icc_set_sgi.

Due to the value of val will write the register, so
change val type to unsigned long. Source code please
reference as below:

*((volatile unsigned int *)((void *)gic_base + GIC_DIST_SOFTINT)) = val;

+++ inter-core-comm.c
@@ -203,7 +203,7 @@ void icc_block_free(unsigned long block)
 
 void icc_set_sgi(unsigned int core_mask, unsigned int hw_irq)
 {
-	unsigned long val;
+	unsigned int val;
 
 	if(hw_irq > 15) {
 		printf ("Interrupt id num: %u is not valid, SGI[0 - 15]\n", hw_irq);
